+++ src/lil_pwny/loggers.py
@@ -42,6 +42,10 @@ class StdoutLogger:
             message = f'CUSTOM_MATCH: \n' \
                       f'    ACCOUNT: {message.get("username").lower()} HASH: {message.get("hash")} PASSWORD: {message.get("plaintext_password")} OBFUSCATED: {message.get("obfuscated")}'
             mes_type = 'CUSTOM'
+        if notify_type == "username":
+            message = f'USERNAME_MATCH: \n' \
+                      f'    ACCOUNT: {message.get("username").lower()} HASH: {message.get("hash")} PASSWORD: {message.get("plaintext_password")} OBFUSCATED: {message.get("obfuscated")}'
+            mes_type = 'USERNAME'
         if notify_type == "duplicate":
             message = 'DUPLICATE: \n' \
                       f'    ACCOUNTS: {message.get("users")} HASH: {message.get("hash")} OBFUSCATED: {message.get("obfuscated")}'
@@ -114,6 +118,12 @@ class StdoutLogger:
                 base_color = Fore.YELLOW
                 high_color = Fore.YELLOW
                 key_color = Fore.YELLOW
+                style = Style.NORMAL
+                mes_type = '!'
+            elif mes_type == "USERNAME":
+                base_color = Fore.BLUE
+                high_color = Fore.BLUE
+                key_color = Fore.BLUE
                 style = Style.NORMAL
                 mes_type = '!'
 
@@ -195,6 +205,9 @@ class JSONLogger(Logger):
         self.notify_format = logging.Formatter(
             '{"localtime": "%(asctime)s", "level": "NOTIFY", "source": "%(name)s", "match_type": "%(type)s", '
             '"detection_data": %(message)s}')
+        self.success_format = logging.Formatter(
+            '{"localtime": "%(asctime)s", "level": "SUCCESS", "source": "%(name)s", '
+            '"detection_data": %(message)s}')
         self.info_format = logging.Formatter(
             '{"localtime": "%(asctime)s", "level": "%(levelname)s", "source": "%(name)s", "message":'
             ' %(message)s}')
@@ -215,6 +228,11 @@ class JSONLogger(Logger):
     def log(self, level: str, log_data: str or Dict, **kwargs):
         if level.upper() == 'NOTIFY':
             self.handler.setFormatter(self.notify_format)
+            self.logger.info(
+                json.dumps(log_data, cls=EnhancedJSONEncoder),
+                extra={'type': kwargs.get('notify_type', '')})
+        elif level.upper() == 'SUCCESS':
+            self.handler.setFormatter(self.success_format)
             self.logger.info(
                 json.dumps(log_data, cls=EnhancedJSONEncoder),
                 extra={'type': kwargs.get('notify_type', '')})